+++ ymp3/views/__init__.py
@@ -1,0 +1,2 @@
+import generic
+import api_v1
